+++ src/components/Dialog/Dialog.tsx
@@ -1,18 +1,26 @@
-import React from 'react';
+import React, { useRef } from 'react';
 import { useFocusTrap, useLockBodyScroll } from '../../hooks'
 
 import './Dialog.scss'
 import classNames from 'classnames';
 import { UUI } from '../../utils/uui';
+import { useClickAway } from 'react-use';
 
 export interface BaseDialogProps {
   /**
    * Toggles the visibility of the overlay and its children.
    * This prop is required because the component is controlled.
    */
-  open: boolean,
-  focusTrap?: boolean,
-  lockBodyScroll?: boolean,
+  open: boolean
+  /**
+   * Callback invoked when user click outside of content view.
+   */
+  onDismiss?: () => void
+  focusTrap?: boolean
+  /**
+   * Wether lock scrolling on the body element while Drawer is active
+   */
+  lockBodyScroll?: boolean
   /**
    * The content of this dialog
    */
@@ -31,6 +39,12 @@ export const Dialog = UUI.FunctionComponent({
   const { Root, Backdrop, Container, Content } = nodes
 
   useLockBodyScroll(props.open && !!props.lockBodyScroll)
+
+  const containerRef = useRef<any>(null)
+  useClickAway(containerRef, () => {
+    props.open && props.onDismiss && props.onDismiss()
+  })
+
   const ref = useFocusTrap(props.open && !!props.focusTrap)
   return (
     <Root>
@@ -43,6 +57,7 @@ export const Dialog = UUI.FunctionComponent({
         tabIndex={-1}
       >
         <Container
+          ref={containerRef}
           role="dialog"
           className={classNames(['u-bg-white u-rounded u-p-4'], {
             'is-visible': props.open